ai-services/requirements.txt
fastapi
uvicorn
pydantic
langchain-core
langchain-google-genai
python-dotenv
opik
PyJWT